// apps/core/src/seeds/activity-type.enum.ts
export enum ActivityTypeEnum {
  LIKE = 'LIKE',
  COMMENT_LIKE = 'COMMENT_LIKE',
  RECEIVE_LIKE = 'RECEIVE_LIKE',
  SHARE = 'SHARE',
  COMMENT = 'COMMENT',
  CREATE_POLL = 'CREATE_POLL',
  REPORTING = 'REPORTING',
  BUY_ANALYTICS = 'BUY_ANALYTICS', // TODO
  COMPLETE_PROFILE = 'COMPLETE_PROFILE',
  SHARE_APP = 'SHARE_APP',
  POLL_VOTE = 'POLL_VOTE',
  DAILY_ACTIVE_USER = 'DAILY_ACTIVE_USER',
  SAVE_POLL = 'SAVE_POLL',
  REPOST = 'REPOST',
  FIRST_SIGNUP = 'FIRST_SIGNUP',
}